+++ backend/mora/triggers/internal/http_trigger.py
@@ -1,18 +1,21 @@
 # SPDX-FileCopyrightText: 2019-2020 Magenta ApS
 # SPDX-License-Identifier: MPL-2.0
-
 import asyncio
-from structlog import get_logger
 from functools import partial
-from typing import Dict, List
+from typing import Dict
+from typing import List
 
 import aiohttp
-from mora import config
-from mora.async_util import async_session, async_to_sync, in_separate_thread
-from mora.triggers import Trigger
-from os2mo_http_trigger_protocol import MOTriggerPayload, MOTriggerRegister
-from pydantic import parse_obj_as
 from fastapi.encoders import jsonable_encoder
+from mora import config
+from mora.async_util import async_session
+from mora.async_util import async_to_sync
+from mora.async_util import in_separate_thread
+from mora.triggers import Trigger
+from os2mo_http_trigger_protocol import MOTriggerPayload
+from os2mo_http_trigger_protocol import MOTriggerRegister
+from pydantic import parse_obj_as
+from structlog import get_logger
 
 
 logger = get_logger()
@@ -43,16 +46,23 @@ async def http_sender(trigger_url: str, trigger_dict: dict, timeout: int):
 
             None is returned if no exceptions occur.
     """
-    logger.debug("http_sender called", trigger_url=trigger_url,
-                 trigger_dict=trigger_dict, timeout=timeout)
-    timeout = aiohttp.ClientTimeout(total=timeout)
+    logger.debug(
+        "http_sender called",
+        trigger_url=trigger_url,
+        trigger_dict=trigger_dict,
+        timeout=timeout,
+    )
+    client_timeout = aiohttp.ClientTimeout(total=timeout)
     async with async_session() as session:
         # TODO: Consider changing trigger_dict to MOTriggerPayload throughout
         payload = jsonable_encoder(MOTriggerPayload(**trigger_dict).dict())
-        async with session.post(trigger_url, timeout=timeout, json=payload) as response:
+        async with session.post(
+            trigger_url, timeout=client_timeout, json=payload
+        ) as response:
             payload = await response.json()
-            logger.debug("http_sender received",
-                         payload=payload, trigger_url=trigger_url)
+            logger.debug(
+                "http_sender received", payload=payload, trigger_url=trigger_url
+            )
             if response.status != 200:
                 raise HTTPTriggerException(payload["detail"])
             return payload
@@ -91,20 +101,23 @@ async def fetch_endpoint_trigger(
     full_url = endpoint + "/triggers"
 
     logger.debug("Sending request to", full_url=full_url)
-    timeout = aiohttp.ClientTimeout(total=timeout)
+    client_timeout = aiohttp.ClientTimeout(total=timeout)
     try:
-        async with session.get(full_url, timeout=timeout) as response:
+        async with session.get(full_url, timeout=client_timeout) as response:
             try:
                 trigger_configuration = parse_obj_as(
                     List[MOTriggerRegister], await response.json()
                 )
                 return trigger_configuration
-            except aiohttp.client_exceptions.ContentTypeError:
-                logger.error("Unable to parse response from (not JSON?)",
-                             full_url=full_url)
+            except aiohttp.client_exceptions.ContentTypeError as exc:
+                logger.error(
+                    "Unable to parse response from (not JSON?)", full_url=full_url
+                )
                 logger.error(reponse_text=await response.text())
-    except asyncio.exceptions.TimeoutError:
+                raise exc
+    except asyncio.exceptions.TimeoutError as exc:
         logger.error("Timeout while connecting to", full_url=full_url)
+        raise exc
 
 
 @in_separate_thread
@@ -138,14 +151,14 @@ def register(app) -> bool:
     """Register triggers for what the http trigger handlers need.
 
     This method:
-    * Checks the configuration of the module
+    * Checks the configuration of the module.
     * Fetches /triggers on all configured http trigger handlers.
     * Registers the http_sender trigger for all the requested events.
     """
     settings = config.get_settings()
     endpoints = settings.http_endpoints
     if not endpoints:
-        logger.warning("Module enabled without endpoints!")
+        logger.debug("HTTP Triggers has no endpoints!")
         return False
 
     # Timeout for fetching endpoint trigger configurations (per endpoint)
@@ -156,8 +169,9 @@ def register(app) -> bool:
 
     # Fetch configured triggers for all endpoints
     endpoint_trigger_dict = fetch_endpoint_triggers(endpoints, fetch_trigger_timeout)
-    logger.debug("Got endpoint_trigger_dict",
-                 endpoint_trigger_dict=endpoint_trigger_dict)
+    logger.debug(
+        "Got endpoint_trigger_dict", endpoint_trigger_dict=endpoint_trigger_dict
+    )
 
     # Register http_sender for all the events found.
     for endpoint, trigger_configuration in endpoint_trigger_dict.items():